rename utility back to ffi (preparing to merge with master...)
