+++ pywincffi/kernel32/process.py
@@ -16,7 +16,7 @@ documentation for the constant names and their purpose:
 """
 
 import six
-from pywincffi.core.utility import Library, ffi, input_check, error_check
+from pywincffi.core.ffi import Library, ffi, input_check, error_check
 
 kernel32 = Library.load("kernel32")
