+++ tests/test_core/test_utility.py
@@ -10,7 +10,7 @@ from cffi import FFI
 from six.moves import builtins
 
 import pywincffi
-from pywincffi.core.utility import (
+from pywincffi.core.ffi import (
     NON_ZERO, Library, new_ffi, ffi, input_check, error_check)
 from pywincffi.core.testutil import TestCase
 from pywincffi.exceptions import (
